Move UserSession to the IdentityManagement service
