+++ src/Exception/TokenNotFoundException.php
@@ -2,7 +2,7 @@
 
 namespace Lullabot\Mpx\Exception;
 
-use Lullabot\Mpx\UserSession;
+use Lullabot\Mpx\Service\IdentityManagement\UserSession;
 
 /**
  * Exception thrown when a token is not found in the token cache.
@@ -12,9 +12,9 @@ class TokenNotFoundException extends \RuntimeException
     /**
      * TokenNotFoundException constructor.
      *
-     * @param UserSession     $user     The user the token was not found for.
-     * @param int             $code     The exception code.
-     * @param \Throwable|null $previous The previous exception, if available.
+     * @param \Lullabot\Mpx\Service\IdentityManagement\UserSession $user     The user the token was not found for.
+     * @param int                                                  $code     The exception code.
+     * @param \Throwable|null                                      $previous The previous exception, if available.
      */
     public function __construct(UserSession $user, int $code = 0, \Throwable $previous = null)
     {
