+++ src/Service/IdentityManagement/AuthenticatedClient.php
@@ -2,11 +2,10 @@
 
 namespace Lullabot\Mpx\Service\IdentityManagement;
 
-use GuzzleHttp\Promise\PromiseInterface;
-use Lullabot\Mpx\Client;
-use Lullabot\Mpx\UserSession;
 use GuzzleHttp\ClientInterface;
 use GuzzleHttp\Promise\Promise;
+use GuzzleHttp\Promise\PromiseInterface;
+use Lullabot\Mpx\Client;
 use Lullabot\Mpx\Exception\ClientException;
 use Psr\Http\Message\RequestInterface;
 
@@ -15,7 +14,7 @@ class AuthenticatedClient implements ClientInterface
     /**
      * The user to establish a client for.
      *
-     * @var \Lullabot\Mpx\UserSession
+     * @var \Lullabot\Mpx\Service\IdentityManagement\UserSession
      */
     protected $user;
 
@@ -32,8 +31,8 @@ class AuthenticatedClient implements ClientInterface
      * Note that the authentication is not actually established until
      * acquireToken is called.
      *
-     * @param \Lullabot\Mpx\Client      $client The client used to access MPX.
-     * @param \Lullabot\Mpx\UserSession $user   The user associated with this client.
+     * @param \Lullabot\Mpx\Client                                 $client The client used to access MPX.
+     * @param \Lullabot\Mpx\Service\IdentityManagement\UserSession $user   The user associated with this client.
      */
     public function __construct(
         Client $client,
