+++ tests/src/Unit/Service/AccessManagement/ResolveAllUrlsTest.php
@@ -5,10 +5,10 @@ namespace Lullabot\Mpx\Tests\Unit\Service\AccessManagement;
 use Cache\Adapter\PHPArray\ArrayCachePool;
 use Lullabot\Mpx\Service\AccessManagement\ResolveAllUrls;
 use Lullabot\Mpx\Service\IdentityManagement\AuthenticatedClient;
+use Lullabot\Mpx\Service\IdentityManagement\UserSession;
 use Lullabot\Mpx\Tests\JsonResponse;
 use Lullabot\Mpx\Tests\MockClientTrait;
 use Lullabot\Mpx\TokenCachePool;
-use Lullabot\Mpx\UserSession;
 use PHPUnit\Framework\TestCase;
 use Psr\Log\NullLogger;
 use Symfony\Component\Lock\StoreInterface;
